task 1234 plan sketch seems ok. tried that task 4 plan sketch grounded plans don't work. tried that task 1 plan sketch grounded plans don't work
task4 ground plan:

Crack the egg egg-0 into bowl-0 and discard the eggshell.
Crack the egg egg-2 into bowl-0 and discard the eggshell.
Pour the entire sugar-1 in measuring-cup-6 into bowl-0.
Pour the entire flour-1 in measuring-cup-1 into bowl-0.
Put the butter butter-0 in the measuring cup into bowl-0.
Pour the entire baking-powder-0 in measuring-cup-2 into bowl-0.
Mix the ingredients in bowl-0 with the mixer mixer-0 to create a new mixture mixture-0 with all the ingredients fused together.
Close the oven door if it's open, and preheat the oven oven-0 at 350 degrees Fahrenheit, the temperature at which to bake cakes.
Pour the entire mixture mixture-0 from bowl-0 into pan-0.
Put the pan-0 in the oven-0.
Set the amount of time to bake a cake and press 'start' on oven-0 to begin baking.
Crack the egg egg-4 into bowl-1 and discard the eggshell.
Crack the egg egg-5 into bowl-1 and discard the eggshell.
Separate the egg egg-4 into an egg yolk, now called egg-4 and in the container bowl-1, and egg whites called egg-6 and in the container bowl-0.
Separate the egg egg-5 into an egg yolk, now called egg-5 and in the container bowl-1, and egg whites called egg-6 and in the container bowl-0. # First error here. egg-6 is duplicate from prev step; should use a different hypothetical egg object.
Beat the egg egg-6 with the mixer mixer-0 in the bowl-0. # Second error here. The next four steps should be putting in bowl-1, with the egg yolk.
Pour the entire sugar-0 in measuring-cup-3 into bowl-0.
Pour the entire flour-0 in measuring-cup-0 into bowl-0.
Put the butter butter-1 in the measuring cup into bowl-0.
Mix the ingredients in bowl-0 with the mixer mixer-0 to create a new mixture mixture-0 with all the ingredients fused together. # Third error here. The egg whites should not be mixed.
Fold the egg whites egg-6 in bowl-1 into the contents of bowl-0, using the spatula-0. # Fourth error here. The egg whites are mixed.
Close the oven door if it's open, and preheat the oven oven-0 at 375 degrees Fahrenheit, the temperature at which to bake souffles.
Pour the entire mixture mixture-0 from bowl-0 into pan-1.
Put the pan-1 in the oven-0. # Fourth error here. The oven is already occupied from baking the cake and not removing the pan.
Set the amount of time to bake a souffle and press 'start' on oven-0 to begin baking. # Fifth error here. The oven cannot hold the pan.